--- app/testCard/RestaurantInfo.tsx
@@ -1,17 +1,0 @@
-//this is a struct that has info for a specific Restaurant
-//should be used throughout program same way
-
-export interface RestaurantInfo {
-
-    restaurant_id: string;
-    name: string;
-    address: string;
-    reviewCount: number;
-    image_url: string;
-    rating: number;
-    categories: string[];
-    price: number;
-    distance: number;
-    location: string;
-    closing_times: string[];
-}
