+++ components/RenderPage.tsx
